return nil on missing lifecycle action
Redis HGETALL returns an empty list when accessing a missing key. The calling
code in this case is expecting fetchInstanceLifecycleAction() to return nil
in case of a missing action.

Missing action may happen when an instance posts against /terminations/<id>
after imploding. In that case it never received a termination request, so
there is no action to fetch.